Add authenticate. Protecting API Routes

+++ API/app.js
@@ -3,6 +3,8 @@ const app = express();
 const { mongoose } = require('./db/mongoose');
 // Load in the mongoose models
 const { List, Task, User } = require('./db/models');
+const jwt = require('jsonwebtoken');
+
 
 //#region MIDDLEWARE
 
@@ -25,6 +27,26 @@ app.use(function (req, res, next) {
 	next();
 });
 
+
+// check it the request has a valid jwt secret
+let authenticate = (req, res, next) => {
+	let token = req.header('x-access-token');
+
+	// verify the JWT
+	jwt.verify(token, User.getJWTSecret(), (err, decoded) => {
+		if (err) {
+			// jwt is invalid
+			res.status(401).send(err);
+		}
+		else {
+			// jwt is valid
+			req.user_id = decoded._id;
+
+			next();
+		}
+	});
+}
+
 // verify refresh token middleware
 let verifySession = (req, res, next) => {
 	// get refresh token from the request header
@@ -45,9 +67,9 @@ let verifySession = (req, res, next) => {
 		req.user_id = user._id;
 		req.userObject = user;
 		req.refreshToken = refreshToken;
-		
+
 		let isSessionValid = false;
-		
+
 		user.sessions.forEach((session) => {
 			if (session.token === refreshToken) {
 				// check if the session is expired
@@ -81,10 +103,14 @@ let verifySession = (req, res, next) => {
  * GET /lists
  * Purpose: Get all lists
  */
-app.get('/lists', (req, res) => {
+app.get('/lists', authenticate, (req, res) => {
 	// We want to return an array of all the lists that belong to the authenticated user 
-	List.find().then((lists) => {
+	List.find({
+		_userId: req.user_id
+	}).then((lists) => {
 		res.send(lists);
+	}).catch((e) => {
+		res.send(e);
 	});
 })
 
@@ -92,13 +118,14 @@ app.get('/lists', (req, res) => {
  * POST /lists
  * Purpose: Create a list
  */
-app.post('/lists', (req, res) => {
+app.post('/lists', authenticate, (req, res) => {
 	// We want to create a new list and return the new list document back to the user (which includes the id)
 	// The list information (fields) will be passed in via the JSON request body
 	let title = req.body.title;
 
 	let newList = new List({
-		title
+		title,
+		_userId: req.user_id
 	});
 
 	newList.save().then((listDoc) => {
@@ -111,9 +138,9 @@ app.post('/lists', (req, res) => {
  * PATCH /lists/:id
  * Purpose: Update a specified list
  */
-app.patch('/lists/:id', (req, res) => {
+app.patch('/lists/:id', authenticate, (req, res) => {
 	// We want to update the specified list (list document with id in the URL) with the new values specified in the JSON body of the request
-	List.findOneAndUpdate({ _id: req.params.id }, {
+	List.findOneAndUpdate({ _id: req.params.id, _userId: req.user_id }, {
 		$set: req.body
 	}).then(() => {
 		res.sendStatus(200);
@@ -124,12 +151,16 @@ app.patch('/lists/:id', (req, res) => {
  * DELETE /lists/:id
  * Purpose: Delete a list
  */
-app.delete('/lists/:id', (req, res) => {
+app.delete('/lists/:id', authenticate, (req, res) => {
 	// We want to delete the specified list (document with id in the URL)
 	List.findOneAndRemove({
-		_id: req.params.id
+		_id: req.params.id,
+		_userId: req.user_id
 	}).then((removedListDoc) => {
 		res.send(removedListDoc);
+
+		// delete all the tasks
+		deleteTasksByListId(removedListDoc._id);
 	})
 })
 
@@ -137,7 +168,7 @@ app.delete('/lists/:id', (req, res) => {
  * GET /lists/:listId/tasks
  * Purpose: Get all tasks in a specific list
  */
-app.get('/lists/:listId/tasks', (req, res) => {
+app.get('/lists/:listId/tasks', authenticate, (req, res) => {
 	// We want to return all tasks that belong to a specific list (specified by listId)
 	Task.find({
 		_listId: req.params.listId
@@ -150,14 +181,32 @@ app.get('/lists/:listId/tasks', (req, res) => {
  * POST /lists/:listId/tasks
  * Purpose: Create a new task in a specific list
  */
-app.post('/lists/:listId/tasks', (req, res) => {
+app.post('/lists/:listId/tasks', authenticate, (req, res) => {
 	// We want to create a new task in a list specified by listId
-	let newTask = new Task({
-		title: req.body.title,
-		_listId: req.params.listId
-	});
-	newTask.save().then((newTaskDoc) => {
-		res.send(newTaskDoc);
+
+	List.findOne({
+		_id: req.params.listId,
+		_userId: req.user_id
+	}).then((list) => {
+		if (list) {
+			// list exists
+			return true;
+		}
+
+		return false;
+	}).then((canCreateTask) => {
+		if (canCreateTask) {
+			let newTask = new Task({
+				title: req.body.title,
+				_listId: req.params.listId
+			});
+			newTask.save().then((newTaskDoc) => {
+				res.send(newTaskDoc);
+			})
+		}
+		else {
+			res.sendStatus(404);
+		}
 	})
 })
 
@@ -165,31 +214,66 @@ app.post('/lists/:listId/tasks', (req, res) => {
  * PATCH /lists/:listId/tasks/:taskId
  * Purpose: Update an existing task
  */
-app.patch('/lists/:listId/tasks/:taskId', (req, res) => {
+app.patch('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
 	// We want to update an existing task (specified by taskId)
-	Task.findOneAndUpdate({
-		_id: req.params.taskId,
-		_listId: req.params.listId
-	}, {
-		$set: req.body
-	}
-	).then(() => {
-		res.send({ message: 'Updated successfully.' })
-	})
+
+	List.findOne({
+		_id: req.params.listId,
+		_userId: req.user_id
+	}).then((list) => {
+		if (list) {
+			// list exists
+			return true;
+		}
+
+		return false;
+	}).then((canUpdateTask) => {
+		if (canUpdateTask) {
+			Task.findOneAndUpdate({
+				_id: req.params.taskId,
+				_listId: req.params.listId
+			}, {
+				$set: req.body
+			}
+			).then(() => {
+				res.send({ message: 'Updated successfully.' })
+			}) 
+		}
+		else {
+			res.sendStatus(404);
+		}
+	});
+
+	
 });
 
 /**
  * DELETE /lists/:listId/tasks/:taskId
  * Purpose: Delete a task
  */
-app.delete('/lists/:listId/tasks/:taskId', (req, res) => {
-	Task.findOneAndRemove({
-		_id: req.params.taskId,
-		_listId: req.params.listId
-	}).then((removedTaskDoc) => {
-		res.send(removedTaskDoc);
-	})
-})
+app.delete('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
+
+	List.findOne({
+		_id: req.params.listId,
+		_userId: req.user_id
+	}).then((list) => {
+		if (list) {
+			// list exists
+			return true;
+		}
+
+		return false;
+	}).then((canDeleteTask) => {
+		if (canDeleteTask) {
+			Task.findOneAndRemove({
+				_id: req.params.taskId,
+				_listId: req.params.listId
+			}).then((removedTaskDoc) => {
+				res.send(removedTaskDoc);
+			})
+		}
+	});
+});
 
 
 //#region User
@@ -261,6 +345,15 @@ app.get('/users/me/access-token', verifySession, (req, res) => {
 })
 
 //#endregion
+
+/* HELPER METHOD */
+let deleteTasksByListId = (_listId) => {
+	Task.deleteMany({
+		_listId
+	}).then(() => {
+		console.log("list id : " + _listId + " deleted");
+	});
+}
 
 app.listen(3000, () => {
 	console.log("server is 3000")
